Removed some compiler warnings and updated the amiga makefiles.
Patch submitted by Dirk Stoecker <stoecker@rcs.urz.tu-dresden.de>

[[originally from CVS; cvs2svn created svn r2726]]

+++ client/gui-mui/objecttreeclass.c
@@ -227,7 +227,7 @@ HOOKPROTONH(ObjectTree_Group_Layout, ULONG, Object *obj, struct MUI_LayoutMsg *l
     ret = TRUE;
     break;
   default:
-    ret = MUILM_UNKNOWN;
+    ret = (ULONG) MUILM_UNKNOWN;
   }
   return ret;
 }
